Updated Use of Color note
Rereading the material on 1.4.1, I thought it needed a bit more info on 1.4.11 in order to clarify how both considerations need to be brought to bear when considering state indicators. It's possible that this section could be moved to another part of the understanding document, and benefit from some visual examples.
Also, added "large" to make it "large text" in first paragraph of Understanding.