Bo's Intent and Active components changes
I removed the parenthetical qualifier for LV users in the first sentence and moved it down. This is covered in the rationale for 1.4.3 which is referenced in the very next sentence. I added "control" in parentheses to address Bo's concern.
In Active components, tried to clarify language around boundary to match discussion from working group.