Minor updates for clarity
One edit I didn't do because it's in the graph image and the text that follows it.

"Most of them have good contrast except the green triangles."
The green line appears to be 2.2:1. I think this line is sending a mixed message. Is it ok to have most of the lines right and have a failing line  or does this example fail the SC. I would say its the later. I think we should remove the green line from the example or make it darker passing green.