Removed G139 from Situation B techniques
G139 is about Creating a mechanism that allows users to jump to errors. For pratical purposes it will almost invariably involve a change in focus (otherwise the user would need to hunt for the mechanism) or else will appear on page reload (following an attemtp at submitting a form). As such, it is not the most appropriate technique to site in Status Messages, and I think the Understanding document is better without it.