Update status-messages.html
Removed second failure technique as per email to Alastair and AWK:  "The second is misplaced (and I'm not sure who proposed it initially). My read on visibilitychange is that it is not a state that is set by a developer, so really a developer would still need to set display:none (or whatever) based on the visibilitychange. And most of the methods used to hide content would also hide it from a screen reader. So I plan to remove that from the understanding doc. Would you also please remove it from the spreadsheet? If you are aware of live region information NOT acting in concert with something like display:none, please let me know.