Add paragraph about non-transcribed CAPTCHA audio
Closes #3198

Note: Issue #323, from 2018, does bring up the issue of what generated on the fly audio is (live? prerecorded?) and it wasn’t resolved. It’s not clear to me whether this fits under 1.2.1 (pre-recorded audio only) or 1.2.9 (audio-only live), but I’m not sure it matters as both require an alternative.