Focus Not Obscured: revise uses of 'hidden'
closes #3446

'hidden' is generally used to refer to content that is not presented to any user (e.g., display none) or 'hidden' from the accessibility tree (e.g., aria-hidden=true).  

The current use of 'hidden' in the normative text / understanding doc _mostly_ is meant to mean 'visually hidden/obscured' (though there is an instance of "hidden" in the understanding doc that _was_ referring to the 'completely hidden' definition.

This PR changes instances of 'hidden' that meant 'visually hidden / visually obscured' to use these terms, instead of just 'hidden'.